abdml_msfworks/requirements.txt
absl-py==2.2.0
asttokens==3.0.0
astunparse==1.6.3
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
comm==0.2.2
confluent-kafka==2.8.2
contourpy==1.3.1
convertdate==2.4.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
ephem==4.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
geomet==0.2.1.post1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
holidays==0.69
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
LunarCalendar==0.0.9
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
prophet==1.1.6
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyMeeus==0.5.12
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
stanio==0.5.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
